Organize payments vertically (#225)
* organize payments vertically

* update docs

* docs updates and small changes
